% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_data.R
\name{acs_vars_builder}
\alias{acs_vars_builder}
\title{ACS Variable Builder}
\usage{
acs_vars_builder(
  year,
  dataset = "acs5",
  detailed_tagging = FALSE,
  return = TRUE,
  verbose = FALSE
)
}
\arguments{
\item{year}{Year of ACS data return.}

\item{dataset}{Defaults to "acs5", but can use any valid ACS dataset. 
Currently, this and all my functions are tuned to the ACS dataset. 
Future functionality may include decennial data and PUMS.}

\item{detailed_tagging}{Logical: TRUE provides column tagging kind of ACS var table.}

\item{return}{If FALSE, load variables into env., otherwise return to variable.}

\item{verbose}{Logical parameter to specify whether to produce verbose output.}
}
\value{
Produces two objects in the global environment: ACS.VARS and ACS.TABLES.
ACS.VARS is a master list of all variables, ACS.TABLES is a filtered dataframe
of top-level variables.
}
\description{
A simple function to construct the variable lists returned from the ACS dataset.
Uses tidycensus::load_variables() and saves both the output as a global object,
and a filtered object as a summary file. Primarily internal use.
}
\examples{
\dontrun{
acs_vars_builder()
}
}
