% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_data.R
\name{comparison_helper}
\alias{comparison_helper}
\title{Comparison Helper}
\usage{
comparison_helper(
  df = NULL,
  comparisonDF = NULL,
  comp_type = NULL,
  tableID = NULL,
  stateFilter = NULL,
  verbose = FALSE
)
}
\arguments{
\item{df}{A primary dataset to provide comparisons against.}

\item{comparisonDF}{A geography-level comparison dataset.}

\item{comp_type}{Explicitly call comparison type.}

\item{tableID}{The tableID to reference comparison.}

\item{stateFilter}{FIPS or abbreviation for filtering and selecting comparisons.}

\item{verbose}{Logical parameter to specify whether to produce verbose output.}
}
\value{
Dataframe
}
\description{
Comparison Helper
}
\examples{
\dontrun{
comparison_helper(data,comparisondata,tableID="B02001",stateFilter=17)
}
}
