% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_data.R
\name{create_comparison_data}
\alias{create_comparison_data}
\title{Create Profile (Batch By Geographies)}
\usage{
create_comparison_data(
  geography = NULL,
  profileDataset = NULL,
  year = NULL,
  variables = NULL,
  tableID = NULL,
  coordColName = "sf",
  verbose = FALSE,
  showCall = FALSE,
  geosObject = NULL,
  censusDataset = "acs5",
  test = FALSE
)
}
\arguments{
\item{geography}{Either "us" or "state".}

\item{profileDataset}{Optional dataset to determine state/county for filtering.}

\item{year}{Numeric value specifying year of ACS call.}

\item{variables}{A variables vector.}

\item{tableID}{A tableID vector.}

\item{coordColName}{Default, set to "sf", but can be changed if non-sf object.}

\item{verbose}{Whether to provide verbose output.}

\item{showCall}{Whether get_acs() should have showCall set to TRUE}

\item{geosObject}{Optional geosObject to speed up processing time.}

\item{censusDataset}{The census dataset for calls. Defaults to `acs5`.}

\item{test}{Internal: for testing purposes.}
}
\value{
A deep, nested list. Structure is list1 > regionInfo,data.
regionInfo is a tibble with identifiers, and data contains three tibbles:
df, dfCount, dfNoSummary. See create_profile for more information.
To access individual datasets, use this structure: object$data[[n]]$profileTable.
}
\description{
This function relies on create_profile_batch for its internal logic, but 
is used primarily to create comparison profile objects for whole geographies.
This is useful when rendering displayTable() and a larger geography 
comparison is desired.
}
\examples{
\dontrun{
create_comparison_data(geography="state",
year=2021,variables = profile_variables,tableID = profile_tableID)
create_comparison_data(geography="us",year=2021,
variables = profile_variables,tableID = profile_tableID)
}
}
