% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_data.R
\name{entropyIndex}
\alias{entropyIndex}
\title{Entropy Index}
\usage{
entropyIndex(
  data = NULL,
  dataFormat = "long",
  majorityVar = NULL,
  minorityVar = NULL,
  geography = "tract",
  wideCols = NULL,
  longCol = "pct",
  filterSummary = FALSE,
  filterSummaryLevels = "root",
  tableID = NULL,
  variables = NULL,
  filterAddress = NULL,
  filterRadius = NULL,
  state = NULL,
  county = NULL,
  tract = NULL,
  block_group = NULL,
  year = NULL,
  return = FALSE,
  dataset_main = "acs",
  dataset_sub = "acs5",
  dataset_last = NULL,
  censusVars = NULL,
  geosObject = NULL,
  verbose = FALSE
)

entropyIndex(
  data = NULL,
  dataFormat = "long",
  majorityVar = NULL,
  minorityVar = NULL,
  geography = "tract",
  wideCols = NULL,
  longCol = "pct",
  filterSummary = FALSE,
  filterSummaryLevels = "root",
  tableID = NULL,
  variables = NULL,
  filterAddress = NULL,
  filterRadius = NULL,
  state = NULL,
  county = NULL,
  tract = NULL,
  block_group = NULL,
  year = NULL,
  return = FALSE,
  dataset_main = "acs",
  dataset_sub = "acs5",
  dataset_last = NULL,
  censusVars = NULL,
  geosObject = NULL,
  verbose = FALSE
)
}
\arguments{
\item{data}{A data object for which to estimate entropy.}

\item{geography}{Defaults to "tract," but must match data object.}

\item{wideCols}{Specified columns for calculating entropy in wide data.}

\item{longCol}{Specified columns for calculating entropy in long data.}

\item{filterSummary}{Logical parameter to specify whether to filter out summary levels (typically _001 and therefore "root").}

\item{filterSummaryLevels}{Explicit description of lowest type denoting summary level. Also excludes lower levels.#'}

\item{tableID}{TableID specification for calculating entropy.}

\item{variables}{Variable specification for calculating entropy.}

\item{filterAddress}{For data calls, a filtered area specification.}

\item{filterRadius}{For data calls, a filtered area specification.}

\item{state}{Input (abb. or FIPS) of state for search.}

\item{county}{Input (abb. or FIPS) of county for search.}

\item{tract}{Input (abb. or FIPS) of tract for search.}

\item{block_group}{Input (abb. or FIPS) of block group for search.}

\item{year}{Year for data call.}

\item{return}{Logical parameter. If TRUE, return value only. Otherwise return formatted string.}

\item{dataset_main}{Selection parameters for get_census_variables (e.g. "acs")}

\item{dataset_sub}{Selection parameters for get_census_variables (e.g. "acs5")}

\item{dataset_last}{Selection parameters for get_census_variables (e.g. "cprofile")}

\item{censusVars}{Passthrough object to bypass get_census_variables}

\item{geosObject}{Passthrough object to bypass get_geocode_radius}

\item{verbose}{Logical parameter to specify whether to produce verbose output.}

\item{dataType}{Can choose "long", "wide" or "vector" depending on data object type.}

\item{dissimilarityValue}{For dissimilarity index, selecting minority group.
Numeric value corresponding to variable.}

\item{dissimilarityValueB}{For exposure/isolation index, selecting minority
group. Numeric value corresponding to variable.}
}
\value{
dataframe

dataframe
}
\description{
A statistical function to estimate diversity / segregation in datasets.

A statistical function to estimate diversity / segregation in datasets.
}
\details{
Logic developed from
https://www2.census.gov/programs-surveys/demo/about/housing-patterns/multigroup_entropy.pdf
and Scientific Study of Religion - 2016 - Dougherty - Congregational Diversity
and Attendance in a Mainline Protestant-2024-02-19-13-18.pdf OR
https://onlinelibrary.wiley.com/doi/epdf/10.1111/j.1468-5906.2008.00390.x."EI"
is built from the second, while MGEI is built from the former. Basically, the
EI signifies how much diversity exists in a community (e.g., Census Tract).
The Multigroup index of the whole metro area examines how much segregation
exist between areas in a larger area. It doesn't account for the diversity of
the whole, but the integration of the areas. So for example, Individual census
tracts may have a high entropy score (0.8), signifying lots of diversity. But
it may have a low metro score (.05) suggesting that diversity is evenly spread
throughout the larger area. If the metro MGEI was higher, it would signify
more segregation between diverse areas.
}
\examples{
\dontrun{
entropyIndex(data=NULL,tableID = "B11012",variables =
c(1:4),year=2022,verbose=TRUE,filterAddress = v,filterRadius = 1)
}
\dontrun{
entropyIndex(data=NULL,tableID = "B11012",variables =
c(1:4),year=2022,verbose=TRUE,filterAddress = v,filterRadius = 1)
}
}
