% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_geo.R
\name{geo_road_helper}
\alias{geo_road_helper}
\title{Geo: Road Helper}
\usage{
geo_road_helper(df = NULL, verbose = FALSE)
}
\arguments{
\item{df}{The road sf object.}

\item{verbose}{A logical parameter to specify verbose output.}
}
\value{
A dataframe with "suffix" column added, for filtering use later.
}
\description{
A geo utility to assist filtering roads which do not meet certain criteria, 
to avoid bogging down tmap "plot" maps with too much text.
}
\details{
First, strips out all directional pre/suffixes. Then, returns a modified
dataframe with added suffix column. Can use this to filter later.
}
\examples{
\dontrun{
  roads <- tigris::roads(state=geo$states,county = geo$counties,year=2021,filter_by=bbox)
  roads <- geo_road_helper(roads)
  }
}
