% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_geo.R
\name{geo_var_builder}
\alias{geo_var_builder}
\title{Geography Variable Builder}
\usage{
geo_var_builder(
  geography = c("all"),
  try = "download",
  state = NULL,
  county = NULL,
  geosObject = NULL,
  verbose = FALSE,
  test = FALSE
)
}
\arguments{
\item{geography}{Specifies the kind of geography to call.
Options: "state","county","tract","block". Defaults to "all".}

\item{try}{Parameter to specify whether to try local file read first, before downloading.}

\item{state}{Filter by county name or COUNTYFP}

\item{county}{Filter by state name or STATEFP}

\item{geosObject}{Optional geosObject object to speed up geo processing.}

\item{verbose}{Logical param to provide feedback.}

\item{test}{Logical param to bypass feedback and messages during testing.}
}
\value{
Returns global objects if does not exist.
}
\description{
Checks if geospatial objects exist for referencing census data. If not, get them.
}
\examples{
\dontrun{
geo_var_builder()
geo_var_builder(geography="state")
geo_var_builder(geography="county")
geo_var_builder(geography="tract")
geo_var_builder(geography="block")
}

}
