% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_geo.R
\name{geocoder}
\alias{geocoder}
\title{Geocoder}
\usage{
geocoder(
  address,
  year = 2020,
  service = "census",
  batch = FALSE,
  verbose = FALSE
)
}
\arguments{
\item{address}{Address entry in STREET, CITY, STATE ZIP format.}

\item{year}{Year to get census geocoding. See census info for year specs.}

\item{service}{Whether to use census geocoder or OSM.}

\item{batch}{Internal: for geocoder_batch.}

\item{verbose}{Logical param to provide feedback.}
}
\value{
Dataframe with coordinates.
}
\description{
Using Census Geocoder
}
\examples{
\dontrun{
#Using census data
geocoder(address="350 Fifth Avenue New York, NY 10118",year=2020,service='census')

#Using OSM data
geocoder(address="350 Fifth Avenue New York, NY 10118",service='OSM')
}

}
