% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_geo.R
\name{geocoder_batch}
\alias{geocoder_batch}
\title{Geocoder Batch}
\usage{
geocoder_batch(
  addressList,
  addressCol = "address",
  start = NULL,
  limit = NULL,
  na.rm = TRUE,
  batch = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{addressList}{List of addresses to geocode.}

\item{addressCol}{Name or number of address column.}

\item{start}{If needed, a starting ID to begin batch processing.}

\item{limit}{If needed, an ending ID to end batch processing.}

\item{na.rm}{To remove NA. NOT YET IMPLEMENTED}

\item{verbose}{Logical param to provide feedback.}
}
\value{
A modified dataframe including coordinates.
}
\description{
Geocoder Batch
}
\examples{
\dontrun{
geocoder_batch(addressList)
geocoder_batch(addressList,start=10,limit=50)
}

}
