% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_data.R
\name{get_census_variables}
\alias{get_census_variables}
\title{Get Census Variables}
\usage{
get_census_variables(
  year = NULL,
  dataset_main = NULL,
  dataset_sub = NULL,
  dataset_last = NULL,
  dataset_extra = NULL,
  detailed_tagging = FALSE,
  directory = FALSE,
  raw = FALSE,
  verbose = FALSE
)
}
\arguments{
\item{year}{Year for variable draw}

\item{dataset_main}{Main dataset parameter (e.g., 'acs' or 'dec')}

\item{dataset_sub}{Secondary dataset parameter (e.g., 'acs5')}

\item{dataset_last}{Tertiary dataset parameter (e.g., 'cprofile' or 'subject')}

\item{detailed_tagging}{Logical, to flag additional data for dataframe.}

\item{directory}{Logical, to grab list of all datasets available in census API}

\item{verbose}{Logical parameter to signal verbose output.}
}
\value{
A dataframe
}
\description{
Sends a JSON request to the Census API to capture variables for use in 
other functionality.
}
\examples{
\dontrun{
get_census_variables(year=2022, dataset_main="acs", dataset_sub="acs5")
}
}
