% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_geo.R
\name{get_geocode_radius}
\alias{get_geocode_radius}
\title{Get a Geocoded Radius}
\usage{
get_geocode_radius(
  filterAddress = NULL,
  filterRadius = NULL,
  filterByGeoType = NULL,
  filterByGeoValue = NULL,
  state = NULL,
  county = NULL,
  geoidLookup = NULL,
  geography = NULL,
  radiusOnly = FALSE,
  geocodeOnly = FALSE,
  fipsOnly = FALSE,
  profile = FALSE,
  coords = NULL,
  intersectOverlap = 0.1,
  year = NULL,
  test = FALSE,
  verbose = FALSE,
  geosObject = NULL
)
}
\arguments{
\item{filterAddress}{Address for geocoding. Should be in "Street, City, State Zip"
format if possible. Names are also optional, but consistency of lookup is not
guaranteed.}

\item{filterRadius}{An integer describing the width of the radius in miles.}

\item{filterByGeoType}{An irregular geo type to get a smaller overlapping set
of tracts, block_groups or other geography from. Options are currently
"metro", "place","combined_statistical_areas". E.g., Find all tracts in Chicago (place).}

\item{filterByGeoValue}{A value to find object for filtering. Either NAME or GEOID.}

\item{state}{Input (abb. or FIPS) of state for search.}

\item{county}{Input (abb. or FIPS) of county for search.}

\item{geoidLookup}{Lookup geography by GEOID.}

\item{geography}{Typically passed by other functions, used for capi()
request to specify geography. May be either "us","state", "county","tract", or "block group".}

\item{radiusOnly}{Set to TRUE to return geocoded radius only, and not
provide any additional ACS data requests.}

\item{geocodeOnly}{Set to TRUE to return the geocode of the address only
as an object.}

\item{fipsOnly}{Logical param to return only FIPS from radius.}

\item{profile}{A profile data object.}

\item{coords}{Internal: passing already found coordinates to function.}

\item{intersectOverlap}{A proportion of minimum required overlap from radius
and unit area.}

\item{year}{Year for data call}

\item{test}{Internal: Parameter to allow testing passthroughs [depreciated]}

\item{verbose}{Internal: Parameter for printing feedback.}

\item{geosObject}{Optional geosObject object to speed up geo processing.}
}
\value{
Returns either 1) a filtered sf object from capi() call; 2) an sf
object with point coordinates; 3) an sf object with radius coordinates;
or 4) a tmap map with radius overlay.
}
\description{
A function to geocode a supplied address, and then return a specified list
with various coordinate information.
}
\details{
Designed to work with censusprofiler functions, and using OpenStreetMap /
US Census data, supplied addresses are geocoded. Resulting coordinates are
processed, including radius buffers. Parameters are returned to the
originating function, where capi() is called. Alternatively, it can be used
to generate simply a geocoded address, or a geospatial radius.
}
\examples{
\dontrun{
#Get full ACS dataset
get_geocode_radius(filterAddress="350 Fifth Avenue New York, NY 10118",
filterRadius=1,geometry=TRUE)

#Get the radius only.
get_geocode_radius(filterAddress="350 Fifth Avenue New York, NY 10118",
filterRadius=1,radiusOnly=TRUE)

#Get a simple geocode of an address.
get_geocode_radius(filterAddress="350 Fifth Avenue New York, NY 10118",
filterRadius=1,geocodeOnly=TRUE)
}
}
