% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapper.R
\name{mapper}
\alias{mapper}
\title{mappeR}
\usage{
mapper(
  mapDF = NULL,
  tableID = NULL,
  variable = NULL,
  variableSummary = NULL,
  year = NULL,
  geography = NULL,
  filterAddress = NULL,
  filterRadius = NULL,
  coords = NULL,
  tract = NULL,
  county = NULL,
  state = NULL,
  geoidLookup = NULL,
  dispPerc = FALSE,
  LegendTitle = NULL,
  MapTitle = NULL,
  radiusOnly = FALSE,
  areaOnly = FALSE,
  neighbors = FALSE,
  neighbor_depth = 1,
  alpha = 0.5,
  interactive = FALSE,
  geosObject = NULL,
  markers = NULL,
  dispRoads = TRUE,
  dispWater = TRUE,
  dispPlaces = TRUE,
  road_resolution = 1,
  dispRails = FALSE,
  dataset_main = NULL,
  dataset_sub = NULL,
  dataset_last = NULL,
  censusVars = NULL,
  verbose = FALSE,
  test = FALSE,
  st = NULL
)
}
\arguments{
\item{mapDF}{A census profiler object.}

\item{tableID}{Variable base, prior to the underscore: i.e. "B01001"  ->>
TODO Necessary???}

\item{variable}{Since maps are not ideal for displaying a range of variables,
the single variable number to display.}

\item{variableSummary}{Summary variable against which to calculate the
percentage.}

\item{year}{Year for capi().}

\item{geography}{Either "us","state","county","tract", or "block group".}

\item{filterAddress}{Address for querying a set of geographies as a radius
around a location.}

\item{filterRadius}{The radius in miles.}

\item{coords}{Parameter to pass coordinates to function for quicker
geolocation.}

\item{tract}{A list of specific tracts to filter.}

\item{county}{County by name or FIPS code.}

\item{state}{State by name or FIPS code.}

\item{geoidLookup}{Lookup geography by GEOID.}

\item{dispPerc}{Set to TRUE to calculate percentage based on summary
variable.}

\item{LegendTitle}{Set for custom title.}

\item{MapTitle}{Text value to set map title.}

\item{radiusOnly}{To display the tracts/counties/states and the radius on a
map.}

\item{areaOnly}{To display the tracts/counties/states within the area
specified by geography.}

\item{neighbors}{Logical parameter to specify whether neighboring geos should be called.}

\item{neighbor_depth}{Integer value of depth of neighbor rings.}

\item{alpha}{Set alpha value of map.}

\item{interactive}{Parameter for setting tmap_mode to either `interactive` or
`plot.`}

\item{geosObject}{Optional geosObject to speed up processing time.}

\item{markers}{A geos object containing sf files for roads, water, places or
rails. Intended to speed up repeated uses of mapper, as in RMD files.}

\item{dispRoads}{Logical: whether to display roads on non-interactive map.}

\item{dispWater}{Logical: whether to display water bodies on non-interactive
map.}

\item{dispPlaces}{Logical: whether to display places on non-interactive map.}

\item{road_resolution}{Possible values: c(1,2,3,4). Value of 1 displays all roads
and names; value of 2 shows all roads, but only potentially major road names; 
value of 3 shows all roads and no names; value of 4 shows and names only 
potentially major roads; value of 5 shows only potential major roads and no names.}

\item{dispRails}{Logical: whether to display railroads on non-interactive
map.}

\item{dataset_main}{Selection parameters for get_census_variables (e.g. "acs")}

\item{dataset_sub}{Selection parameters for get_census_variables (e.g. "acs5")}

\item{dataset_last}{Selection parameters for get_census_variables (e.g. "cprofile")}

\item{censusVars}{Passthrough object to bypass get_census_variables}

\item{verbose}{Logical param to provide feedback.}

\item{test}{Internal: Logical param to bypass feedback and provide limited
data for testing.}

\item{st}{Internal variable passed from other functions to provide
consistency on timestamp.}
}
\value{
Returns a filtered tmap map, detailing either estimates or
  percentages of variables requested.
}
\description{
mappeR takes ACS data pulled from capi() and creates a map displaying
estimates, either as simple counts, or percentages based on summary
variables.
}
\details{
Logics include: 1) a call with variables + filterAddress (capi() call)
[variable,variableSummary,filterAddress,filterRadius]; 2) a call with
variables + filterAddress + df included
[mapDF,variable,variableSummary,filterAddress,filterRadius]; 3) a call
without variables + filterAddress displaying area by geography
[filterAddress,filterRadius,areaOnly]; 4) a call without variables +
geography specified (e.g.: state=17, geography="county")
[state/county/tract,areaOnly]; 5) a call without variables + filterAddress
and geography, but only radius displayed
[filterAddress,filterRadius,radiusOnly]
}
\examples{
\dontrun{
Create a map of a selected region based on a variable.
mapper(mapDF = NULL,tableID = "B02001",variable = "B02001_002",variableSummary =
"B02001_001",geography = "tract",year = 2022,dispPerc = FALSE,
MapTitle = "Selected Census Tracts",LegendTitle =
"Census Tracts",filterAddress = address,filterRadius = 1,alpha =
0.1,interactive = FALSE,geosObject = NULL,verbose = FALSE)

Create a map of an entire geography, using a variable.
mapper(mapDF = NULL,tableID = "B02001",variable = "B02001_002",variableSummary =
"B02001_001",geography = "county",year = 2022,state = 17,county
= 043,dispPerc = FALSE,MapTitle = "Selected Census Tracts",LegendTitle =
"Census Tracts",alpha = 0.1,interactive = FALSE,geosObject = NULL,
verbose = FALSE)
}

}
