% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_data.R
\name{profile_batch}
\alias{profile_batch}
\title{Profile Batch}
\usage{
profile_batch(
  batch_name = NULL,
  name_column = NULL,
  address_column = NULL,
  year = NULL,
  tableID = NULL,
  variables = NULL,
  geography = NULL,
  addressList = NULL,
  filterRadius = NULL,
  filterByGeoType = NULL,
  filterByGeoValue = NULL,
  filterSummary = FALSE,
  filterSummaryLevels = "root",
  dataset_main = "acs",
  dataset_sub = "acs5",
  dataset_last = NULL,
  censusVars = NULL,
  state = NULL,
  county = NULL,
  tract = NULL,
  block_group = NULL,
  metro = NULL,
  ggr = NULL,
  geosObject = NULL,
  simpleReturn = FALSE,
  test = FALSE,
  fast = FALSE,
  verbose = FALSE,
  st = NULL
)
}
\arguments{
\item{batch_name}{Master name for profile batch.}

\item{name_column}{Column with titles or names of entries. If blank, will assume column #1.}

\item{address_column}{Column with titles or names of entries. If blank, will assume column #2.}

\item{year}{Year for data selection.}

\item{tableID}{Specification for concept, or group: e.g., "B01001"}

\item{variables}{A vector of all variables requested.}

\item{geography}{Geography specification: e.g.: tract, county, state.}

\item{addressList}{A dataframe containing addresses for lookup. If
coordinates are supplied, profile_batch will bypass geocode functions,
thereby speeding up processing.}

\item{filterRadius}{A numeric value specifying the radius in miles around the address.}

\item{filterByGeoType}{An irregular geo type to get a smaller overlapping set
of tracts, block_groups or other geography from. Options are currently
"metro", "place","combined_statistical_areas". E.g., Find all tracts in Chicago (place).}

\item{filterByGeoValue}{A value to find object for filtering. Either NAME or GEOID.}

\item{filterSummary}{Logical parameter to specify whether to filter out summary levels (typically _001 and therefore "root").}

\item{filterSummaryLevels}{Explicit description of lowest type denoting summary level. Also excludes lower levels.#' @param state Input (abb. or FIPS) of state for search.}

\item{dataset_main}{Selection parameters for get_census_variables (e.g. "acs")}

\item{dataset_sub}{Selection parameters for get_census_variables (e.g. "acs5")}

\item{dataset_last}{Selection parameters for get_census_variables (e.g. "cprofile")}

\item{censusVars}{Passthrough object to bypass get_census_variables}

\item{state}{Input (abb. or FIPS) of state for search.}

\item{county}{Input (abb. or FIPS) of county for search.}

\item{tract}{Input (abb. or FIPS) of tract for search.}

\item{block_group}{Input (abb. or FIPS) of block group for search.}

\item{metro}{Input (abb. or FIPS) of metropolitan statistical area for search.}

\item{ggr}{Internal: to pass a get_geocode_radius() object to function.}

\item{geosObject}{Optional, attach geos object to simplify geo processes.}

\item{simpleReturn}{Param to return raw data, not formatted.}

\item{test}{Internal parameter for testing suite.}

\item{fast}{Internal parameter for stat table (capi())}

\item{verbose}{Logical parameter to specify whether to produce verbose output.}

\item{st}{Internal parameter to provide timestamp consistency.}
}
\value{
A nested list with profile objects.
}
\description{
Profile Batch
}
\examples{
\dontrun{
profile_batch <- profile_batch(batch_name = "TEST",addressList =
test,name_column = "Name",address_column = "Address",
tableID="B02001",geography="tract",verbose=T,censusVars =
CV,year=2022,filterRadius = 1)
}

}
