% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profiler.R
\name{profiler}
\alias{profiler}
\title{ProfileR}
\usage{
profiler(
  name = NULL,
  year = NULL,
  dataset_main = "acs",
  dataset_sub = "acs5",
  dataset_last = NULL,
  censusVars = NULL,
  tableID = NULL,
  variables = NULL,
  geography = NULL,
  filterAddress = NULL,
  filterRadius = NULL,
  coords = NULL,
  filterByGeoType = NULL,
  filterByGeoValue = NULL,
  filterSummary = FALSE,
  filterSummaryLevels = "root",
  state = NULL,
  county = NULL,
  tract = NULL,
  block_group = NULL,
  metro = NULL,
  place = NULL,
  ggr = NULL,
  geosObject = NULL,
  simpleReturn = FALSE,
  test = FALSE,
  fast = FALSE,
  verbose = FALSE,
  st = NULL
)
}
\arguments{
\item{name}{User-supplied name for profile.}

\item{year}{Year for data selection.}

\item{dataset_main}{Selection parameters for get_census_variables (e.g. "acs")}

\item{dataset_sub}{Selection parameters for get_census_variables (e.g. "acs5")}

\item{dataset_last}{Selection parameters for get_census_variables (e.g. "cprofile")}

\item{censusVars}{Passthrough object to bypass get_census_variables}

\item{tableID}{Specification for concept, or group: e.g., "B01001"}

\item{variables}{A vector of all variables requested.}

\item{geography}{Geography specification: e.g.: tract, county, state.}

\item{filterAddress}{An address input used to generate a radius around, for filtering data.}

\item{filterRadius}{A numeric value specifying the radius in miles around the address.}

\item{coords}{An sf coordinate object for get_geocode_radius.}

\item{filterByGeoType}{An irregular geo type to get a smaller overlapping set
of tracts, block_groups or other geography from. Options are currently
"metro", "place","combined_statistical_areas". E.g., Find all tracts in Chicago (place).}

\item{filterByGeoValue}{A value to find object for filtering. Either NAME or GEOID.}

\item{filterSummary}{Logical parameter to specify whether to filter out summary levels (typically _001 and therefore "root").}

\item{filterSummaryLevels}{Explicit description of lowest type denoting summary level. Also excludes lower levels.#' @param state Input (abb. or FIPS) of state for search.}

\item{state}{Input (abb. or FIPS) of state for search.}

\item{county}{Input (abb. or FIPS) of county for search.}

\item{tract}{Input (abb. or FIPS) of tract for search.}

\item{block_group}{Input (abb. or FIPS) of block group for search.}

\item{metro}{Input (abb. or FIPS) of metropolitan statistical area for search.}

\item{place}{Input (abb. or FIPS) of place for search.}

\item{ggr}{Internal: to pass a get_geocode_radius() object to function.}

\item{geosObject}{Optional, attach geos object to simplify geo processes.}

\item{simpleReturn}{Param to return raw data, not formatted.}

\item{test}{Internal parameter for testing suite.}

\item{fast}{Internal parameter for stat table (capi())}

\item{verbose}{Logical parameter to specify whether to produce verbose output.}

\item{st}{Internal parameter to provide timestamp consistency.}
}
\value{
data.frame
}
\description{
A wrapper for capi() with additional parameters for profile creation.
}
\examples{
\dontrun{
make_profile(x)
}
}
