% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_data.R
\name{segregationMeasures}
\alias{segregationMeasures}
\title{Segration Measures}
\usage{
segregationMeasures(
  data = NULL,
  dataFormat = "long",
  majorityVar = NULL,
  minorityVar = NULL,
  geography = "tract",
  wideCols = NULL,
  longCol = "pct",
  filterSummary = FALSE,
  filterSummaryLevels = "root",
  tableID = NULL,
  variables = NULL,
  filterAddress = NULL,
  filterRadius = NULL,
  state = NULL,
  county = NULL,
  tract = NULL,
  block_group = NULL,
  year = NULL,
  return = FALSE,
  tips = FALSE,
  dataset_main = "acs",
  dataset_sub = "acs5",
  dataset_last = NULL,
  censusVars = NULL,
  geosObject = NULL,
  verbose = FALSE
)
}
\arguments{
\item{data}{A data object for which to estimate entropy.}

\item{dataFormat}{Can choose "long", "wide" or "vector" depending on data object type.}

\item{majorityVar}{For dissimilarity index, selecting minority group.
Numeric value corresponding to variable.}

\item{minorityVar}{For exposure/isolation index, selecting minority
group. Numeric value corresponding to variable.}

\item{geography}{Defaults to "tract," but must match data object.}

\item{wideCols}{Specified columns for calculating entropy in wide data.}

\item{longCol}{Specified columns for calculating entropy in long data.}

\item{filterSummary}{Logical parameter to specify whether to filter out summary levels (typically _001 and therefore "root").}

\item{filterSummaryLevels}{Explicit description of lowest type denoting summary level. Also excludes lower levels.#'}

\item{tableID}{TableID specification for calculating entropy.}

\item{variables}{Variable specification for calculating entropy.}

\item{filterAddress}{For data calls, a filtered area specification.}

\item{filterRadius}{For data calls, a filtered area specification.}

\item{state}{Input (abb. or FIPS) of state for search.}

\item{county}{Input (abb. or FIPS) of county for search.}

\item{tract}{Input (abb. or FIPS) of tract for search.}

\item{block_group}{Input (abb. or FIPS) of block group for search.}

\item{year}{Year for data call.}

\item{return}{Logical parameter. If TRUE, return value only. Otherwise return formatted string.}

\item{tips}{Logical parameter specifying whether to include helpful tips.}

\item{dataset_main}{Selection parameters for get_census_variables (e.g. "acs")}

\item{dataset_sub}{Selection parameters for get_census_variables (e.g. "acs5")}

\item{dataset_last}{Selection parameters for get_census_variables (e.g. "cprofile")}

\item{censusVars}{Passthrough object to bypass get_census_variables}

\item{geosObject}{Passthrough object to bypass get_geocode_radius}

\item{verbose}{Logical parameter to specify whether to produce verbose output.}
}
\value{
dataframe
}
\description{
A statistical function to estimate diversity / segregation in datasets.
}
\examples{
\dontrun{
entropyIndex(data=NULL,tableID = "B11012",variables =
c(1:4),year=2022,verbose=TRUE,filterAddress = v,filterRadius = 1)
}
}
