% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_data.R
\name{set_api_key}
\alias{set_api_key}
\title{Set API Key}
\usage{
set_api_key(key = NULL, test = FALSE)
}
\arguments{
\item{key}{The API key obtained for use with US Census Data API.}

\item{test}{Internal, for testing purposes.}
}
\value{
A global environmental variable.
}
\description{
Sets the API key for the US Census API (see
https://api.census.gov/data/key_signup.html) into the global environment for
reuse with censusprofiler.
}
\examples{
\dontrun{
set_api_key(APIKEYGOESHERE)
}

}
