% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_geo.R
\name{spatial_helper}
\alias{spatial_helper}
\title{Spatial Helper}
\usage{
spatial_helper(
  df,
  geography = "tract",
  state = NULL,
  county = NULL,
  geosObject = NULL,
  test = FALSE,
  verbose = FALSE
)
}
\arguments{
\item{df}{Dataobject with GEOID column.}

\item{geography}{Param specifying the geographical designation for census information.}

\item{state}{Filter geo_var_builder by state to speed things up.}

\item{county}{Filter geo_var_builder by county to speed things up.}

\item{geosObject}{Optional geosObject to speed up processing time.}

\item{test}{Internal: Logical param to bypass feedback and provide limited data for testing.}

\item{verbose}{Logical param to provide feedback.}
}
\value{
An sf object
}
\description{
Internal helper function to create spatial objects by merging geoid with geo_tracts data.
}
\examples{
\dontrun{
spatial_helper(data)
}
}
