% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_data.R
\name{speedtest}
\alias{speedtest}
\title{Speedtest}
\usage{
speedtest(x, y)
}
\arguments{
\item{x}{First comparison object}

\item{y}{Second comparison object}
}
\value{
data.frame with time comparisons.
}
\description{
Speedtest
}
\examples{
\dontrun{
speedtest(x,y)
}

}
