% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_data.R
\name{stat_helper}
\alias{stat_helper}
\title{Utility: Stat Helper}
\usage{
stat_helper(
  data,
  statTable = NULL,
  variables = NULL,
  entryNum = 1,
  zThresh = 1.5,
  zType = "pct",
  tableID = NULL,
  typeFilter = NULL,
  variable = NULL,
  dataType = 4,
  year = 2022,
  dataset_main = "acs",
  dataset_sub = "acs5",
  dataset_last = NULL,
  censusVars = NULL,
  verbose = FALSE
)
}
\arguments{
\item{data}{A profile data object created using create_profile().}

\item{statTable}{A dataframe of census tracts or blocks created with stat_table_builder. Input can receive either object or file path.}

\item{variables}{List of tableID names to test tableID input.}

\item{entryNum}{For looping. If a single use, value is `1`. Otherwise, dynamically populate.}

\item{zThresh}{The threshold for z-score divergence.}

\item{zType}{Whether we're examining relative (`pct`) or absolute (`estimate`).}

\item{tableID}{The tableID value for filtering.}

\item{typeFilter}{Numeric value to filter type levels in analysis.}

\item{variable}{The variable value for filtering.}

\item{dataType}{Allows us to specify whether this is a df, dfCount, or dfNoSummary table.}

\item{year}{Year for data call.}

\item{dataset_main}{Selection parameters for get_census_variables (e.g. "acs")}

\item{dataset_sub}{Selection parameters for get_census_variables (e.g. "acs5")}

\item{dataset_last}{Selection parameters for get_census_variables (e.g. "cprofile")}

\item{censusVars}{Passthrough object to bypass get_census_variables}

\item{verbose}{Set to `TRUE` to print output on rmarkdown (or other) report.}
}
\value{
Either a mutated dataframe, or text output describing significqant z-score divergences.
}
\description{
A function to a) mutate a dataframe of profile variables with Z-scores, and/or b) 
to test whether individual entries are significant on Rmd reports.
}
\examples{
\dontrun{
stat_helper(df,entryNum=1,zThresh=1.5,zType="pct",tableID="B01001",dataType="df",verbose=TRUE)
}
}
