% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_data.R
\name{stat_table_builder}
\alias{stat_table_builder}
\title{Utility: Stat Table Builder}
\usage{
stat_table_builder(
  year = NULL,
  data = NULL,
  summary_table = FALSE,
  master_list = FALSE,
  compiler = FALSE,
  tableID = NULL,
  variables = NULL,
  geography = "tract",
  test = FALSE,
  geosObject = NULL,
  saveProgress = FALSE,
  stateStart = NULL,
  dataset_main = "acs",
  dataset_sub = "acs5",
  dataset_last = NULL,
  censusVars = NULL,
  verbose = FALSE
)
}
\arguments{
\item{year}{Year value for variable selection.}

\item{data}{A data object generated by "master_list" set to TRUE, to create summary_table.}

\item{summary_table}{Set to `TRUE` to create basic distribution statistics.}

\item{master_list}{Set to `TRUE` to download census tracts with variable list provided.}

\item{compiler}{Set to `TRUE` to stitch downloaded master list files into a single master file.}

\item{tableID}{profileList object to pass to ACS call.}

\item{variables}{Variable list (vector) object to pass to ACS call.}

\item{geography}{Options to get "block group","tract","county" or "state".}

\item{test}{Testing variable. Uses a limited data call for internal testing.}

\item{geosObject}{Optional geosObject to speed up processing time.}

\item{saveProgress}{Logical parameter to save individual downloaded files to
/data/ folder to reduce runtime and memory load.}

\item{stateStart}{Related to saveProgress, if the process is interrupted, a 
state FIPS code to begin the process at.}

\item{dataset_main}{Selection parameters for get_census_variables (e.g. "acs")}

\item{dataset_sub}{Selection parameters for get_census_variables (e.g. "acs5")}

\item{dataset_last}{Selection parameters for get_census_variables (e.g. "cprofile")}

\item{censusVars}{Passthrough object to bypass get_census_variables}

\item{verbose}{Logical parameter to specify whether to produce verbose output.}
}
\value{
A dataframe.
}
\description{
A utility function for generating a dataframe object of all census tracts in 
the United States, with variables from a variable list, Additionally, can calculate
distribution statistics.
}
\examples{
\dontrun{
# Run summary statistics.
stat_table_builder(data=profile_all_summary_stats,
summary_table=TRUE) 
# Download all census tracts.
stat_table_builder(master_list=TRUE) 
}
}
