% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_data.R
\name{type_data}
\alias{type_data}
\title{Type Data}
\usage{
type_data(dataset, return = FALSE)
}
\arguments{
\item{dataset}{A dataset to run type_data() on.}

\item{return}{Whether to return the dataframe (TRUE) or type (FALSE)}
}
\value{
numeric value
}
\description{
Type Data
}
\examples{
\dontrun{
type_data(x)
}
}
