% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_data.R
\name{variable_builder}
\alias{variable_builder}
\title{Variable Builder}
\usage{
variable_builder(
  tableID = NULL,
  censusVars = NULL,
  varStartNum = NULL,
  varEndNum = NULL,
  varSummaryNum = NULL,
  varArray = NULL
)
}
\arguments{
\item{tableID}{Variable base, prior to the underscore.}

\item{censusVars}{Reference to variable table.}

\item{varStartNum}{Start number of variables.}

\item{varEndNum}{End number of variables.}

\item{varSummaryNum}{Summary number.}

\item{varArray}{A manual list if need be.}
}
\value{
Returns either a vector, or a nested list.
}
\description{
A function to loop through and grab variables in a nested list for further
analysis.
}
\examples{
\dontrun{
variable_builder(c("B01001","B01002"),censusVars=ACS.VARS,
varStartNum = c(1,1),varEndNum=c(3,3), varArray=NULL,varSummaryNum = c(1,1))
}
}
