% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_data.R
\name{get_vre_table}
\alias{get_vre_table}
\title{Get VRE Tables}
\usage{
get_vre_table(
  data = NULL,
  year = NULL,
  geography = NULL,
  tableID,
  variableList,
  variableAgg = FALSE,
  state = NULL,
  county = NULL,
  censusDataset = "acs5",
  verbose = FALSE,
  savePath = paste(getwd(), "/data/VRE/", sep = "")
)
}
\arguments{
\item{data}{ACS data passed to function from capi().}

\item{year}{ACS year passed to function from capi().}

\item{geography}{ACS geography passed to function from capi().}

\item{tableID}{ACS variable base passed to function from capi().}

\item{variableList}{ACS variable number(s) passed to function from capi(). Integer or list.}

\item{variableAgg}{Future TODO: make this suitable for a list of variables.}

\item{state}{ACS state (descriptive or FIPS) passed to function from capi().}

\item{county}{ACS county (descriptive or FIPS) passed to function from capi().}

\item{censusDataset}{Can select "acs1", "acsse" (supplemental est), "acs3", "acs5","flows" (migration flows)}

\item{verbose}{Logical parameter to display output}

\item{savePath}{Local save path for VRE tables.}
}
\value{
Returns a dataframe with variance, standard error, and margin of error for aggregate data.
}
\description{
A function to calculate the margin of error based on aggregate ACS data.
}
\details{
The basic logic of this function is derived from the Census Bureau 
(https://www.census.gov/data/academy/webinars/2020/calculating-margins-of-error-acs.html).
Information is stored in VRE tables on the CB FTP website in .csv files. 
This function finds, downloads, extracts, converts the .csv files, and then 
performs statistical calculations on the tables, for the aggregate data.
}
\examples{
\dontrun{
  VREMOE <- get_vre_table(data = df,year = 2021,geography = "tract",
  tableID = "B01001",variableList = c(2,3,3),variableAgg = NULL,state = "IL",
  county = 043)
  }

}
