% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{load_data}
\alias{load_data}
\title{Load Data}
\usage{
load_data(
  load_censusVars = FALSE,
  load_geos = FALSE,
  load_stats = FALSE,
  load_profile_compare = FALSE,
  geography = "tract",
  geo_data = c("state", "county", "tract"),
  dataset_main = "acs",
  dataset_sub = "acs5",
  dataset_last = NULL,
  censusVars = NULL,
  loadToGlobal = TRUE,
  year = 2021,
  tableID = NULL,
  variables = NULL,
  geosObject = NULL,
  test = FALSE,
  verbose = FALSE
)
}
\arguments{
\item{load_censusVars}{Logical param to capture census variables and concepts.}

\item{load_geos}{Logical param to capture geos object.}

\item{load_stats}{Logical param to create stats_object.}

\item{load_profile_compare}{Logical param to create geo profile comparison object.}

\item{geography}{Default geography option for stat_table_builder.}

\item{geo_data}{Default geography options for load_geos.}

\item{dataset_main}{Selection parameters for get_census_variables (e.g. "acs")}

\item{dataset_sub}{Selection parameters for get_census_variables (e.g. "acs5")}

\item{dataset_last}{Selection parameters for get_census_variables (e.g. "cprofile")}

\item{censusVars}{Passthrough object to bypass get_census_variables}

\item{loadToGlobal}{Logical param to save to global environment.}

\item{year}{Default year for data calls.}

\item{tableID}{ProfileList object for stat_table_builder.}

\item{variables}{Variable list for stat_table_builder.}

\item{geosObject}{Optional geosObject to speed up processing time.}

\item{test}{Internal logical parameter to specify testing envir.}

\item{verbose}{Logical param to provide feedback.}
}
\value{
data.frame objects loaded into global environment.
}
\description{
A convenience function that loads functional data, including ACS variables, 
a geos object, stats object, and/or geo profile comparison object. These are 
created and then if selected, loaded into the global environment.
}
\examples{
\dontrun{
load_data(load_acs=TRUE, load_geos=TRUE,load_stats=TRUE,
variables=profile_variables,tableID=profile_tableID)
}
}
