% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_geo.R
\name{map_locations}
\alias{map_locations}
\title{Map Locations from Address List}
\usage{
map_locations(
  addressList,
  secondaryAddressList = NULL,
  state = NULL,
  county = NULL,
  density = FALSE,
  density_col = NULL,
  geosObject = NULL,
  verbose = FALSE,
  pointColor = "steelblue",
  secondaryPointColor = "darkred",
  bgColor = "#FFFFFF",
  borderColor = "#444444"
)
}
\arguments{
\item{addressList}{Either a vector with addresses listed, or a dataframe needing formatting.}

\item{secondaryAddressList}{A comparison set of addresses to plot.}

\item{state}{Parameter to filter by state.}

\item{county}{Parameter to filter by county.}

\item{density}{Logical parameter to display cloropleth instead of dots.}

\item{density_col}{Name of density column in addressList.}

\item{geosObject}{Optional geosObject to speed up processing time.}

\item{verbose}{Logical param to provide feedback.}
}
\value{
TMAP map plot
}
\description{
This function takes a vector or dataframe containing addresses and 
geocodes them, and plots them.
}
\examples{
\dontrun{
map_locations(addressList)
map_locations(addressList,state="IL",county=43)
}
}
