% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_data.R
\name{variable_builder}
\alias{variable_builder}
\title{Variable Builder}
\usage{
variable_builder(
  tableID = NULL,
  varStartNum = NULL,
  varEndNum = NULL,
  varSummaryNum = NULL,
  varArray = NULL,
  year = NULL,
  dataset_main = NULL,
  dataset_sub = NULL,
  dataset_last = NULL,
  censusVars = NULL,
  verbose = FALSE
)
}
\arguments{
\item{tableID}{Variable base, prior to the underscore.}

\item{varStartNum}{Start number of variables.}

\item{varEndNum}{End number of variables.}

\item{varSummaryNum}{Summary number.}

\item{varArray}{A manual list if need be.}

\item{year}{Year for variable draw}

\item{dataset_main}{Main dataset parameter (e.g., 'acs' or 'dec')}

\item{dataset_sub}{Secondary dataset parameter (e.g., 'acs5')}

\item{dataset_last}{Tertiary dataset parameter (e.g., 'cprofile' or 'subject')}

\item{censusVars}{Reference to variable table.}

\item{verbose}{Logical parameter to specify additional output.}
}
\value{
Returns either a vector, or a nested list.
}
\description{
A function to loop through and grab variables in a nested list for further
analysis.
}
\examples{
\dontrun{
variable_builder(c("B01001","B01002"),censusVars=CV.VARS,
varStartNum = c(1,1),varEndNum=c(3,3), varArray=NULL,varSummaryNum = c(1,1))
}
}
